change Position() of the plunger element to return a float instead of an integer
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@3724 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ plunger.h
@@ -243,7 +243,7 @@ public:
    STDMETHOD(put_ScatterVelocity)(/*[in]*/ float newVal);
    STDMETHOD(get_MomentumXfer)(/*[out,retval]*/ float *pVal);
    STDMETHOD(put_MomentumXfer)(/*[in]*/ float newVal);
-   STDMETHOD(Position)(int *pVal); //added for mechanical plunger position
+   STDMETHOD(Position)(float *pVal); //added for mechanical plunger position
    STDMETHOD(MotionDevice)(int *pVal);
    STDMETHOD(get_ReflectionEnabled)(/*[out, retval]*/ VARIANT_BOOL *pVal);
    STDMETHOD(put_ReflectionEnabled)(/*[in]*/ VARIANT_BOOL newVal);
